Switch from useBuiltIns "usage" -> "entry"
Prior to this change:

- Bundle size 490626 bytes (124504 bytes gzipped)

After:

- Bundle size 487306 bytes (122293 bytes gzipped)

A little surprised it isn't more, to be honest. When I double-check
this in source-map-explorer I don't see much difference in the core-js
size.

Incidentally, I think this *should* obviate the need for the workaround
implemented in aa401deff4584e31dd55dce, but when I look at the build
dist/client/polyfill.js it looks untransformed to me... May have to do
some further digging.