Try to force useBuiltIns transform to work
Inspecting the dist/ folder shows no transform happening at all, so I
wondered if it was being skipped on entry files and moved the `import`
into the main app file instead. Of course, that's pointless because the
main app file is also an entry point... As an experiment, I moved it
into a non-entry file, with the same result. My next step is going to be
to try tweaking the @babel/preset-env targets to see if I can force a
transform.