Update hooks to work from both branches
On "master" branch (the top level worktree) they do nothing, except for
checking that they are up-to-date.

On the "content" branch (assumed to be a worktree checkout nested under
the main repo at "content/) we check to make sure we're not amending
already-indexed content, and additionally tag content deploys with a
local timestamp.