refactor: implement (experimental) virtual nodes
This is just a WIP that I am going to rollback because it is getting out
of hand. I just want it in the history to reflect the direction of this
exploration.

The basic idea is to allow production functions to accept AST nodes or
"virtual" nodes, and return AST nodes or "virtual" nodes. The purpose of
a virtual node is to temporarily represent something that isn't yet an
AST node itself but will be used as input when constructing one (for
example, a string that came from a token, or an array of nodes that will
become children of another node).

In this way, everything is a node (VirtualNode | NonVirtualNode) instead
of what we previously had (ASTNode | string | undefined | Array<...>).

We want productions to construct real AST nodes that may use virtual
nodes as input (transforming them as opposed to incorporating them
verbatim). Only at the exit of the parser (ie. the `parse()` method) do
we enforce that the node is non-virtual.

However, as you can see from this commit, things get massively
unwieldy very fast. The input to the production functions is basically
`unknown`, which means we have to make heavy use of `asserts` guards,
and these soon become too big to write. For example a combination like
`sequence -> plus -> t` will produce this structure:

    {
      kind: 'VIRTUAL_ARRAY',
      value: [
        {
          kind: 'VIRTUAL_ARRAY',
          value: {
            kind: 'VIRTUAL_STRING',
            value: 'contents'
          }
        },
        // ...
      ]
    }

and "unwrapping" these virtual nodes via assert guards is both
computationally expensive (especially with the nested loops),
repetitive, and probably offers a false sense of security (because the
checks are necessarily incomplete because we don't want to do a
recursive verification that ends up re-traversing the entire AST).

Compared to this "any" is probably better.

I still want to explore ways to get somek safety back though. That might
involve some shallow checks in production functions. I'll see if there
is a way to reduce the amount of boilerplate there. If we do shallow
checks at every level (discretionary, I know) then we should be
relatively sound as a whole. If that works, there might be some way to
tag productions as having been checked and thus make things stricter and
more robust.