fix: squash one remaining type error
I am not super happy about this being marked as a "Node" as it isn't
really. It only every appears as a child of another node, but I need it
here in order TS to accept my grammar, in which one of the rules returns
a `plus()` collection of arguments... Not sure if this is a legit
solution or not but going with it for now to keep the type checker
output clean.