feat: add benchmark:graphql task
So that we can easily compare parser perf against the reference
implementation.

Current numbers are appalling:

- reference "graphql" package: 36.516826ms
- this package prior to implementing memoization: 963.942678ms
- this package at current HEAD: 1046.281615ms

So even though memoization makes it worse, I think we need it in order
to guarantee linear parse time. I expect that with some optimization we
can get lower, but I don't ever expect to reach full parity because we
are competing against a hand-coded recursive descent parser that is
always going to be leaner than our highly dynamic generated parser. I
haven't profiled yet, but I am guessing it's actually going to be the
lexer that is the slowest part, but we'll see.