fix: Hide background bleed on Chrome
Seeing this only on Chrome (not Safari or Firefox), and only above
a certain screen width. The right-hand side (and not the left)
radial-gradient is bleeding into the border zone of the box (even though
there is no border). I suspect some kind of rounding error bug in the
renderer. Whatever the cause, I can stop the leakage with
background-clip.

https://developer.mozilla.org/en-US/docs/Web/CSS/background-clip