chore: force restart
Last commit alone didn't work; it also needed a tweak to my post-receive
hook to force the node_modules to get rebuilt.

Basically, we don't run `yarn install` unless the Node version, the
Yarn version, or the yarn.lock changes. I ran it manually to confirm
that the `yarn install` produces the desired structure, and modified
the hook to re-run if package.json changes too. (For good measure,
added `--frozen-lockfile` and `--non-interactive` -- in addition to the
existing `--offline` -- for more reproducibility, but those aren't at
issue here).