fix: avoid duplicating in-flight pre-flight requests
First version of the cache was very dumb. It would throw away the most
recently added items when it got full, it would only store a result
after the async request succeeded.

Now it re-uses the `LRUCache` class to provide more sensible eviction
semantics, and it stores in-progress results as well so as to avoid the
duplicate pre-flights. Added error handling to deal with failed requests
(cache is cleared for those).