fix: bump baseHeadingLevel down everywhere
I just realized that my headings all jump from h1 to h3 because we have
the outer heading (h1) and then the inner heading, which is calculated
by adding baseHeadingLevel. So, if we have an h1 in our markup, it
becomes an h3 when it should be an h2.

Bad variable name lead me astray here, because "baseHeadingLevel" makes
it sound like the starting heading level, but it's actually an offset to
be added.