fix: avoid wrapping issues on external link icons
Noticed on last blog post:

https://wincent.com/blog/on-the-right-to-disagree

That given DOM like this (as seen in inspector):

    "blah, "
    <a href="..." class="external">
      Something
      ::after
    </a>
    ", blah"

we were getting line break before the comma. This is because the icon
had `display: inline-block` in order to give it size, which causes it to
become untethered from the preceding text even though it is immediately
after it. Obviously very sensitive to text content, which is why I
hadn't seen it before.

As explained here:

https://stackoverflow.com/a/38418279/2103996

we can make leave the `display` of the :after pseudoelement as `inline`,
and use padding to give it size, margin to give it spacing.