fix: React warnings about mismatching dates on server/client
Sadly, I don't think I want to ship this, even though it gets rid of the
warnings:

    Prop `title` did not match. Server: "6/13/2020, 3:19:51 PM" Client: "13/06/2020, 15:19:51"

So, we fix that by using the "Accepts-Language" header to determine the
client preferences on the server, and then we thread that info through
context to make sure that the client rendering uses the exact same
locale.

However, it doesn't work well in practice, even though it does "work".
For example, in my case, my browser sends this header:

    Accept-Language: en-US,en;q=0.9,es;q=0.8,la;q=0.7

So, `date.toLocaleString('en-US')` will actually return:

    "7/4/2020, 5:41:24 PM"

ie. "July, 4th"

On the other hand, `date.toLocaleString()` will return:

    "04/07/2020, 17:41:24"

Because in reality, my date formatting preferences are not reflecting in
my language preferences at all, but rather, deep in the bowels of the
macOS "Language & Region" System Preferences pane, where I have
explicitly opted out of the insane US date system.

So, I don't think this is a workable mechanism, and it's better to just
tolerate the React warning.

Note also, that I was disappointed to find that `Intl.DateTimeFormat`
does not behave the same on NodeJS (the server) as in the browser. At
least, given a minimal example like:

    new Intl.DateTimeFormat('en-US', {
      dateStyle:'medium',
      timeStyle:'long'}
    ).format(date);

Node won't print the time, only the date; the browser prints both. I
could probably hack this to do the same on both by specifying more
options (eg. `hour: 'numeric'` etc), but that seems fiddly and fragile,
and it won't help me with this situation anyway.