chore: add some static files from old hosts
I want to "decommission" the hosts running at www.wincent.com and
secure.wincent.com without breaking any links, if I can help it. They
contain a bunch of PHP pages that do some very simple stuff like
composing a header and a footer and so on. There are only two (basically
dead) pages that handle POST requests.

So, I made this simple Ruby script that runs on the hosts and `curl`s
the pages (and other resources), taking a snapshot of everything. I
normally don't like leaking internal paths like this, but like I said,
the machines are going off the net forever shortly, so it's ok.

The directories that are guarded by .htaccess files all return 403s as
intended, so I haven't included those files in the dump.

As follow up I am going to have to tweak DNS records and SSL
certificates (SAN entries). But then I'll be able to kill off the old
servers.

    require 'fileutils'
    require 'shellwords'

    def escape(word)
      Shellwords.shellescape word
    end

    # Note: we don't have to worry about simlinks over here;
    # both directories contain only directories and regular files
    #
    #       find /var/www/secure.wincent.com/public_html -type l
    #       find /var/www/www.wincent.com/public_html -type l
    #
    %w[
      /var/www/secure.wincent.com/public_html
      /var/www/www.wincent.com/public_html
    ].each do |base|
      puts "Scraping from: #{base}"

      # Note that `Dir.[]` won't return dot-files, so we miss out on:
      #
      #       $ find /var/www/www.wincent.com/public_html -name '.*'
      #       /var/www/www.wincent.com/public_html/.well-known
      #
      # and:
      #
      #       $ find /var/www/secure.wincent.com/public_html -name '.*'
      #       /var/www/secure.wincent.com/public_html/a/products/install/licensees/binary-only/.htaccess
      #       /var/www/secure.wincent.com/public_html/a/products/install/licensees/full-source/.htaccess
      #       /var/www/secure.wincent.com/public_html/a/products/install/licensees/binary+framework/.htaccess
      #       /var/www/secure.wincent.com/public_html/a/products/install/licensees/.htaccess
      #       /var/www/secure.wincent.com/public_html/codegen/.htaccess
      #       /var/www/secure.wincent.com/public_html/.well-known
      #
      Dir["#{base}/**/*"].each do |source|
        host = "#{source}".split(%r{/})[3]
        path = "#{source}".split(%r{/})[5..-1].join('/')
        destination = File.join(__dir__, 'output', source)
        if File.file? source
          if system "curl --silent https://#{host}/#{escape path} -o #{escape destination}"
            puts " OK : #{host}/#{path}"
          else
            puts "FAIL: #{host}/#{path}"
          end
        else
          FileUtils.mkdir_p destination
        end
      end
    end