fix: remove bad index.php
Whoops. I had this formerly harmless file in here, but when we added
`index.php` to our nginx `try_files` directive, it started getting
served up to visitors hitting the root. :facepalm:

So let's kill it, and if anybody goes to https://wincent.com/index.php
they'll just get a 404, which is fine.