fix: improve wrapping of code elements
Chrome seems pretty flaky with `box-decoration-break`. ie. in order to
see the difference between "clone" and "slice" I need to toggle the
border on and off again. Whatever; I think the wrapped spans look better
without borders or rounded corners anyway, so removing those. Also
experimented with bold but it is too heavyweight for my taste.