feat: add support for footnotes
Now that GitHub supports footnotes, I find myself in the habit of using
them there pretty often and therefore wanting to use them here too.

I am not sure if I am fully sold on the design here, but I'm passing in
a DOM identifier so that the `id` attributes in the generated markup can
be unique. One question, for example, is should this field be nullable?
Another is is "DOMIdentifier" the right name? (I wanted to avoid
confusion with GraphQL "id" fields.) Should I be doing more with these
IDs, like using them on title anchor tags? (probably not). Should I
tighten up the schema documentation, where it misleadingly talks of
"prefix"? (As you can see from the link below, the "DOMIdentifier" ends
up being embedded somewhere in the middle of an ID.)

Anyway, all this is possible by setting `docId`, which you can see being
used to construct the anchor names here:

    https://github.com/markdown-it/markdown-it-footnote/blob/337dcc79dfe8b2cb98c32000d8b24691de66105b/index.js#L13

Possible follow-ups: want to check the contrast levels for accessibility
because I eye-balled these and they are probably horrible. May in
general want to do a pass on the styles to make sure they look as pretty
as they should be.