chore: update React 16.13.0 → 17.0.2
Was hoping this would fix some bugs I'm seeing around how click events
are handled (specifically, jump to and back from footnotes is only
working on the first click, or when JS is disabled), but it didn't.
Doesn't hurt to update though, as far as I can tell, so may as well keep
the change.