chore: update Prettier
Although note that we still have an old duplicate copy lying around
because:

    => Found "graphql-relay#prettier@1.19.1"
    info This module exists because "graphql-relay" depends on it.

I don't know/recall why it has it as a non-dev dependency; all its other
deps are under `devDependencies`... looking, seems it is unintended and
was fixed in:

    https://github.com/graphql/graphql-relay-js/commit/0155159f2276fa4c2b194ddfefc85145d9329c2e

but there are too many breaking changes for me to feel all that
enthusiastic about moving from 0.6.0 to the current release (0.9.0):

    https://github.com/graphql/graphql-relay-js/releases

although admittedly, we could get that fix by going just to v0.7.0, I
think, which does look relatively safe. I might try that in the next
commit.