refactor: switch multi invocation to v4 style
As documented here:

    https://github.com/redis/node-redis

But here's the interesting thing about it — it doesn't work for all
Redis commands. So, for example, we can hit a blog permalink page just
fine but not a page using a reverse-sorted index, like the blog index
page.

This shows up as a server crash:

    TypeError: Cannot read property 'posts' of null
        at PostsIndex.render (/Users/wincent/code/masochist/src/client/components/PostsIndex.js:64:35)

Dicking around in the console I established that the API works for some
commands but not others; for example, `ZREVRANGE` is missing, but
`ZCARD` is not:

    in loop ZREVRANGE masochist:6:blog-index [ 0, 2 ]
    command undefined
    in loop ZCARD masochist:6:blog-index []
    command [Function (anonymous)]

So, if I can't find a way to work around this I may just have to
"downgrade" (that is, "upgrade") to the previously working version of
the software.