chore: back out redis "upgrade"
Because it's not worth it.

The `ZREVRANGE` command is not implemented, which breaks things for us.

Note that we can get the same with `ZRANGE` and passing a `REV` option:

    https://github.com/redis/node-redis/issues/1705

but that still leaves headaches elsewhere, where we were using
`ZREVRANGE` with the `WITHSCORES` option. This isn't supported by the
npm package, even though Redis itself supports it:

    https://redis.io/commands/zrange

The suggestion is to switch overt to `ZRANGEBYSCORE` instead, but that
is starting to sound awfully like a hassle.