refactor: replace Gulp with crude script
Script was coded up in haste so is probably riddled with bugs, but it
seems to do what Gulp did without the enormous dependency footprint.

`yarn audit` is a bit happier after this:

```
19 vulnerabilities found - Packages audited: 1167
Severity: 2 Low | 11 Moderate | 6 High
```

Previously, it was:

```
28 vulnerabilities found - Packages audited: 1416
Severity: 2 Low | 17 Moderate | 9 High
```

Most of the remaining issues are in the relay-compiler, of which I am
running a super-old version.

Closes: https://github.com/wincent/masochist/issues/177