refactor: add WIP replacement Redis client
We only need to handle a tiny number of commands:

- GET

(used alone, and...)

- SET
- ZADD
- ZCARD
- ZINCRBY
- ZREM
- ZREMRANGEBYSCORE
- ZREVRANGE

(used in `multi` transactions).

So something crude like this is probably going to be enough. It is
definitely not going to be efficient, but the payloads we are talking
about here are miniscule, so it really won't matter.

Still to do:

- Implement error handling.
- Add `multi` support.
- Address missing features and "TODOs" noted inline (for example, we
  aren't bailing early if we get an incomplete response).
- Add reconnect.
- Maybe write some tests.

Adding this as a .txt file just to show that it's not intended for
actual execution yet, but I do want it in version control in case my
machine dies.