refactor: cut over to WIP Redis client
As noted in parent commit, it has some problems, but it still works well
enough to render the blog index in the local development environment
(this is the only page I've tested so far, and I only tested it once, so
you _know_ this is good).

Split up big file into smaller files so that I can keep things tidy as I
slowly refactor this into something robust.