refactor: make redis client more robust
I tried a few things here but eventually "settled" on this design; that
is to say, I wanted to make the async handling a bit more
robust/idiomatic, so I explored using an async generator, then switched
to using an async iterator. The problem should fix is that when we are
pipelining a bunch of commands and executing them with MULTI + EXEC, we
may get several data payloads back with the responses. In the parent
commit, I could run the little test case multiple times, and my MULTI
request would sometimes get 4 `undefined` responses, or 1, or none.

I think I've fixed that now, and just need to add in error handling
everywhere, but we don't actually expect too many errors as Redis should
pretty reliably return exactly what I'm asking it for.

How does this work? Basically, we want to be able to `await` one line at
a time until we've got everything we need, hence the async iterator.
This ended up being a bit more useful than the generator because we can
pass it around (eg. to parse bulk strings in a subroutine) or into a
recursive call (which we need in order to parse maps, sets, arrays etc).