chore: remove unused Flow types
I stopped using Flow a while ago, and have a pending ticket open for
porting to TypeScript[^1], and another in-progress rewrite anyway[^2]...
so for now these are mostly just noise (at best, you can consider them
as inline documentation that might go stale).

[^1]: https://github.com/wincent/masochist/issues/142
[^2]: https://github.com/wincent/masochist/issues/140 — although note
the ticket is nominally about updating the Relay version, which is way
behind, but which is tantamount to doing a full rewrite anyway because
so much is going to have to change.