fix: don't leak event listeners
Because we were setting up an `on('line')` listener for every command
we ran, we were leaking these quite badly. Wouldn't be a problem if
the clients were short-lived (because when the client is GC'd, the
emitter/listener would go away, because the emitter and the listeners
are all the same object), but the clients live for as long as the server
process runs.