chore: remove express-graphql
I was only using it as a shortcut to getting GraphiQL in dev, but the
schema is basically "finished" at this point and I don't think I'm going
to need to troubleshoot it again (if I'm wrong about that, I can easily
revert this change). In the meantime, killing this is a small step
towards reducing the dependency footprint.