style: switch back to 2-space indent
JS/TS is a 2-spaces kind of language in most places (including most
projects that I've written JS/TS in), but at my previous job we used a
4-space indent (with hard tabs, actually). I was so used to it at the
time that I used it in this project too in order for things not get too
jarring.

Having left that job earlier in the year, now the 4-space indents are
the jarring thing, so let's go back to the old ways again.