feat: add enough code to make NAME recognition work
This is getting unpleasant, but I knew I would get here — the idea is to
feel the pain as a forcing function which drives the implementation to a
more AST-like structure.

That is, instead of having so much logic in the string-based builder,
we'll probably have a layer before that.