feat: add support for "i" flag when parsing atoms
Still need to do the same for ranges and other members in character
classes.

Note this lead me to read the GraphQL spec more carefully and see that
my use of the "i" flag for matching "\u...." was actually wrong — only
"\u" and not "\U" is supported. So I removed the flag; but I'm going to
implement `ignoreCase` support anyway.