feat: add initial implementation of mergeCharacterClasses
Not complete yet, but I've written enough of it that I feel justified
now in coming up with a more robust visitor/transform pattern that deals
with things in an immutable way.

In part, because I want to be able to replace entire nodes now, not just
modify their children. eg.

    {
      kind: 'Alternate',
      children: [
        {kind: 'CharacterClass', ...},
        {kind: 'CharacterClass', ...},
      ],
    }

This node ☝️ should be replaced with a single, merged `CharacterClass`.