refactor: include `id` in NFA nodes
So that I can test these meaningfully. Jest `toEqual` matcher is
checking that objects have the right shape, but it isn't testing for
identity, so that means my test in the previous commit was passing with
this result:

         s1
        /  \
      s0-s2-s4
        \  /
         s3

ie. alternate states s1, s2, s3 all lead to same accept state (s4), but
Jest isn't actually verifying that because all it cares about is that s4
has the right shape. It would also have passed with:

         s1-s4
        /
      s0-s2-s4
        \
         s3-s4

So, let's stick an `id` in there to show it really is the same object.
This will help us later on when (I presume) we'll cut over to using
actual transition tables.

Actual numbers are semi-arbitrary, as they depend on the call graph
shape (ie. as we recurse, we see higher numbers).