chore: dump some WIP code
These are some exploratory hunks that I have had in my work tree over
the last day or so, but I don't think I'm going to go down this path.
Want to commit it and then revert it just to have it in the history, so
that I can get it back if I change my mind.