refactor: add epsilonTo() helper
Because I want to add some helper functions that I can use to address
the TODO in the code about avoiding adding duplicate edges. So, a lot
of these `return` statements and `push` calls are probably going to be
replaced with calls to functions that add edges, including a check for
duplicates.