feat: add `removeEpsilons()`
Note that we really are copying the edges, not just re-using the same
objects. Tests would pass either way, but it seems less likely to cause
confusion if we don't share objects in this way (ie. edges are always
unique; NFA nodes themselves can be shared).

Will add more tests in next commit.