test: add "last" tests for `removeEpsilons()`
Saying "last" here because these are very likely incomplete; these are
just the NFAs that currently exist in the `regExpToNFA()` tests, and I
have some `BUG`/`TODO` comments noted inline where I think we might be
missing coverage. The thing is, they are fairly painful to write or
reason about, which is why I have started with such a minimal set of
examples.