fix: remove assumptions about singular start states
In several places in `regExpToNFA()` we have code that handles multiple
start states and other places where we assume the first node of an NFA
is the only start node.

All the tests pass like that, but it seems best to assume that multiple
start states could exist anywhere, in the name of consistency and
robustness. Tests still pass after the change. If I were a less lazy
person, I would write tests that actually manifest these scenarios.