test: add regression test showing bug in regExpToNFA()
Not a failing test, but it shows the broken NFA that gets produced.

For `/a?b/` we get:

          ε      a
    ->(2)--->(0)--->(1)
        \
         \ ε
          \
           v   b
           (3)--->((4))

There is a missing epsilon transition from state (1) to state (3), which
effectively turns (1) into a dead state, and means that this NFA can
only recognize "b", not "ab".

There may be bugs with other quantifiers too (eg. `/a+b/`, `/a*b/` etc)
so I'll check those out after I've sorted out this problem.