refactor: try setting up epsilon transitions in `minimizeDFA()`
Just checking out my hunch that I mentioned in the previous commit about
needing a unified start transition here, but obviously it doesn't make
any difference: we add the transitions only to have to immediately
remove them (with `removeEpsilons()`) because otherwise `NFAToDFA()`
will throw — we may have to relax that invariant to make this all work.

To be honest, I'm surprised that we weren't already seeing `NFAToDFA()`
throw, because it requires the passed-in NFA to have exactly one start
state (another invariant that I am not sure about).