refactor: make minor improvements to `minimizeDFA()`
Added a new test for ESCAPED_CHARACTER, which was also horribly broken,
then made the necessary fixes to get that working. Changes lead to the
ESCAPED_UNICODE automaton looking slightly less broken than it was, but
it still is broken; I can see the problem now, but will have to wait
until tomorrow to fix it, but in short...

Our `TransitionTable` data type is bad, because it only allows us to
represent DFAs, not all NFAs. This is because we have one row for each
state, and in each row we can map from one input symbol to one target.
An NFA might need to map from one input symbol to multiple targets
though, and we see this come up for sure when we do `transposeTable()`.

A more "classical" transition table would have a fixed-size alphabet and
be represented as a two-dimensional array:

| State | Input "a" | Input "b" |
| ----- | --------- | --------- |
| q0    | q0        | q1        |
| q1    | q0        | q2        |

(etc)

But because we use "Range" objects we have gone with a different layout
to effectively accommodate a variable number of "columns". The question
I need to resolve is whether to allow each "column" to contain specify
multiple target states (makes the "table" a lot more graph-like), or
whether I want to allow a state to have multiple rows in the table
(which makes it harder to look up all outbound transitions because we
can no longer just jump to an array index). I suspect I'll end up going
with the former approach...