feat: implement dead state removal
As these DFA's can still look huge when printed out from their graph
structure form (eg. well over 1000 lines for a seven-state DFA like the
`ESCAPED_UNICODE` one, despite the small number of states), in the test
suite we're going to make assertions using the `TransitionTable`
version. (At this point, I trust the `toTransitionTable` enough that I
feel fine with relying on it in other tests; and in any case, it _has_
to be reliable if this whole thing is going to work at all, because it
is the transition table representation that we'll be using to drive code
generation.)

To make these tests easier to cook up, I made
`stringifyTransitionTable`, to which I can just insert a call like this:

    console.log(stringifyTransitionTable(minimize(/\\u[0-9A-Fa-f]{4}/)));

and then paste the output into the test (after checking it, of course).