test: add more tests for `minimizeDFA()`
Most of these look pretty good now, but note the `xit` ones, which are
disabled because we have some multiple start states in some NFAs. The
constraint we're bumping up against here is that NFA-to-DFA wants to
assume a single start state, but (it seems) that `removeEpsilons()` is
creating machines with multiple start states.