refactor: don't create multiple start states in `removeEpsilons()`
We can fix `minimizeDFA()` by moving (or rather, propagating) our
constraint that requires us to always begin with a single start state.
`removeEpsilons()` won't create multiple start states if it isn't given
any, which means that we can remove the code that was responsible for
creating those states, and `throw` if we ever get something we don't
like.