chore: add new fragments + miscellanea snapshot
Done using:

  (cd .. && git grep -l 'graphql`' | sort | xargs sed -e '/graphql`/,/`/!d') > \
    support/fragments.graphql

Note that it is not 100% fragments (it picks up some duplicates that are
already represented in queries.graphql). I am going to clean this up
(because the crude `sed` extraction includes a bunch of nearby
not-GraphQL text) and rename these files in the next commit to make it
clear what the distinction is (ie. between source and generated
queries).