refactor: use `overflow-wrap: break-word` more extensively
Is this a good idea? This question was asked here:

    https://twitter.com/adamwathan/status/1317915868483956738

and the consensus seems to be it's a bad idea, but only for some
non-English languages, and all my content is in English.

This will stop pages like:

    https://wincent.com/wiki/macOS_%22must_haves%22

which have long URLs visible inside `a` tags from looking like crap on
mobile (where they overflow the width and make the page too wide,
creating a horizontal scroll).