refactor: add `labels` to `Edge` data type
Because I need to store extra info on transitions in order to
distinguish different accept states.

Because we transpose graphs, eliminate and add states, merge states etc
in the course of removing epsilon transitions, converting NFAs to DFAs,
minimizing DFAs, and son on, we can't store these labels in the states
themselves.

Example:

- Take a machine that recognizes "a", another that recognizes "b", and
  create the union of them.
- Convert that union to a DFA and minimize it, and in the end you wind
  up with a start state, an accept state, and two transitions, but
  you've lost the ability to distinguish between the accept states in
  the original machines.

Conversely, if we store the labels in the transitions leading to the
states, we can apply a final step at the end that distinguishes the true
final state based on the transitions leading to it. I have only tested
some simple examples of this so far, but I think the idea will work. If
it doesn't, I may need to totally change the way I do DFA minimization,
hoping that I can do it in a way that preserves the distinctness of the
different accept states. Will cross that bridge when we come to it...