refactor(lexer): extract `equalEdges()` into separate file
It didn't make any sense where it was (in `regExpToNFA`, but not used
there). Potentially I could even fold it in to `removeEpsilons`, the
only place that it is currently used, but I think there's a reasonable
chance that I am going to need to change/reuse it in subsequent commits,
so I am pulling it out.