refactor: add place to store labels in `TransitionTable`
Similar to what we did with the `Edge` component of the `NFA` data
type in 4549825d0, this gives us a place to store labels on edges in
`TransitionTable`.

Doing it this way because it keeps the metadata separate from the core
fields, and the build still typechecks and tests pass with no
modifications. The rationale here is that the "standard" NFA-to-DFA
algorithms, DFA minimization etc, has no concept of labeling states or
transitions, and if I try to cram labels into the core it is just going
to make all that stuff unclear.

What we really need here is an auxiliary layer for lexing that doesn't
interfere too much and I think this fits the bill.

eg. NFAs which care about states and transitions and deal with single
machines vs lexers/scanners which care about working with multiple
machines (and specifically, being able to create unions of them in a way
that doesn't remove our ability to distinguish between the accept states
of the original machines). I don't know if all of this is going to pan
out, but I know it's at least possible in theory because Ragel does this
and a whole lot more (allowing you to embed actions on any transition
and any state, with a large degree of finesse; eg. you can embed actions
on entering states, leaving states, on accept states, on start states
etc). So, if the approach I'm going to try here doesn't work, I will go
back to the drawing board and try something else.