feat: add initial implementation of `union()` helper
Runs, but the labels are getting dropped on the floor for now because I
am not propagating them all the way through. For example, I have taught
`toTransitionTable()` how to copy labels from an NFA, but I haven't
taught `fromTransitionTable()` to do the reverse. I'll do that next, and
there might be other places where I am manually moving or copying edges
(eg. in `NFAToDFA()` and `removeEpsilons()`, for instance), that I'll
have to make sure the labels come across too.