fix: preserve labels via `transposeTable()`
As you can see, this makes the `union()` tests pass in a meaningful way.

The problem was that when we reverse the edges, we weren't also
reversing the source and target states on the labels metadata.

eg. say we had a transition from state 22 to state 23 on "." adding an
`ELLIPSIS` label, when we transposed the NFA, the transition was flipped
to go from state 23 to state 22, but the label was not updated. That is,
it remained as:

    /* 22 */ new Map([['Atom:.', {23: new Set(['ELLIPSIS'])}]]),

but needed to be rewritten as:

    /* 23 */ new Map([['Atom:.', {22: new Set(['ELLIPSIS'])}]]),