refactor: add some rough notes on next steps
Late here, so I don't have time to pursue this right now, but here is a
sketch of some of the things I am thinking about how to leverage these
edge labels.

I want to make the tables easy for the builder to consume, so ultimately
I want distinct states which contain labels that I can read right out of
them to generate the right tokens.