fix: merge labels when going from NFA to DFA
Added a test to highlight the problem here, and found a bug while doing
so. Given two simple machines that have the same transition into an
accept state:

        a      !
  -> 0 ---> 1 ---> 2

        b      !
  -> 3 ---> 4 ---> 5

with corresponding labels, when we union them and minimize we were
getting:

       a,b     !
  -> 0 ---> 1 ---> 2

but only one label was getting carried through to the final transition.

So, I fixed that, but it reveals the deeper problem: here the
minimization is collapsing the states in a way that destroys
information. We know that "!" is carrying us into one of two accept
states, but we don't know which one. What we really want is to prevent
the middle states (1, 4) from merging into a new state; we want this:

              !
     a .-> 1 ---> 2
      /
  -> 0
      \       !
     b .-> 3 ---> 4

I don't think we can reconstruct this information at the end — it's
already too late by then — we need to do something along the way to
preserve it. So far, I don't have a good answer for that: it's all too
easy to either turn our "DFA" into an NFA, or to wind up with something
that is non-minimal. eg. say we make our edge "condition" truly embed
the label instead of just annotating it:

        a     !=A
  -> 0 ---> 1 ---> 2

        b     !=B
  -> 3 ---> 4 ---> 5

When you minimize this, you will get distinct end states, but the
transition isn't _really_ distinct; you actually have an NFA instead of
a DFA:

            ! .-> 3
       a,b   /
  -> 0 ---> 1
             \
            ! .-> 4

One idea I had as to embed a label in the transition leaving from the
start state as well, to show which token we might be recognizing, and
then check on it at the end. ie. If you arrive at the end state with a
label "A" or "B", but you also saw an "A" label at the start, then you
know it's an "A" token. But that fails or at least sucks a bit for even
trivial constructions that share a common prefix; eg.

        +      a      !
  -> 0 ---> 1 ---> 2 ---> 3

        +      b      !
  -> 4 ---> 5 ---> 6 ---> 7

That will minify to:

        +     a,b     !
  -> 0 ---> 1 ---> 2 ---> 3

So the transition leaving the start state will push labels "A" and "B"
onto a stack, and then at the end you will see either "A" and "B" again,
with no way to disambiguate them. And even if you could, you might have
to update and check a costly stack structure to really figure out
which label you'd seen the most... In short, I'm going to have to think
on this some more, and I am not even sure it is entirely possible.