refactor: remove edge-based labels
As noted in the previous commit, these aren't going to handle all the
cases that we need them to handle, so let's get rid of them. The good
news is that we can do all this much, much simpler, so watch out for
subsequent commits.

(The sneak preview, however, is that I realized that because our grammar
can pretty much all be lexed with 1 character of lookahead with only a
couple of special cases that need to be handled specially, like strings
and numbers, we can just generate a minimized DFA for each token, then
connect them all via epsilon transitions, remove the epsilons, and _not_
minimize the result — doing so would make it smaller at the cost of
destroying information, but keeping it as-is gives us something that is
the optimally size machine, for all practical purposes.)